3/4 Passives implemented
-Implemented working passives for 3/4 of the classes
-Drums not yet implemented because I have no ideas for it
-Needs more testing to ensure it works in multiplayer

+++ Assets/UI/HUD/HUD.cs
@@ -154,7 +154,7 @@ namespace Cadenza
         #endregion
         #region Health Bar
 
-        private void OnHealthChanged(int health, ProgressBar bar)
+        private void OnHealthChanged(float health, ProgressBar bar)
         {
             bar.value = health;
         }
